Benchmark using CLI
Remove finished plan

Split plan into 2 separate plans

Was getting kind of big

Implemented first part of plan

Context of Luna was almost exhausted

Fix from review

own function was overkill and buggy for certain names

README for CLI options

Don't always pass RUSTFLAGS via cmdline

When compiling for Rust is that statement true on AMD CPU?
`target-cpu=native` enables AVX2 plus any other useful features
supported by your CPU. Use either `target-cpu=native` or explicit
`+avx2`, not both.

RR compile failed on dioxus-router

Comment

Implemented benchmark part of the plan

Round of bugfixes after Sol review

Remove pointless benchmarks

They were in the `ns` or even `ps` range. Producing annoying red console
output

Remove implemented plans

+++ vda5050-data-types/src/common.rs
@@ -138,8 +138,8 @@ fn parse_common_utc_timestamp(value: &str) -> Option<i64> {
         microsecond *= 10;
     }
 
-    chrono::NaiveDate::from_ymd_opt(year as i32, month as u32, day as u32)?
-        .and_hms_micro_opt(hour as u32, minute as u32, second as u32, microsecond)
+    chrono::NaiveDate::from_ymd_opt(year as i32, month, day)?
+        .and_hms_micro_opt(hour, minute, second, microsecond)
         .map(|datetime| datetime.and_utc().timestamp_micros())
 }
